Remove sentence breaking regex.
Since it doesn't cause a message rejection, it causes issues. Plus "Son"
is a pretty common word to see in Christian texts (for hopefully obvious
reasons). Could re-add this later and change it to
`\bmy (child|son|dear)\b` while replacing it with a general address to
the user (but something better than just 'user').